+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 Version 0.2.1.0
 ---------------
 
-*Unreleased*
+*May 8, 2018*
 
 <https://github.com/mstksg/backprop/releases/tag/v0.2.1.0>