+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+Version 0.2.2.0
+---------------
+
+*Unreleased*
+
+<https://github.com/mstksg/backprop/releases/tag/v0.2.2.0>
+
+*   `Backprop` instance for `BVar`
+
 Version 0.2.1.0
 ---------------